Refine_mementos_distance_visitors

+++ src/main/java/org/example/TurtleMemento.java
@@ -1,4 +1,4 @@
 package org.example;
 
-public record TurtleMemento(double x, double y, double heading) {
+public record TurtleMemento(int x, int y, int heading) {
 }
